+++ src/main/java/de/tum/in/naturals/set/AbstractNatBitSet.java
@@ -23,32 +23,9 @@ import it.unimi.dsi.fastutil.ints.IntIterator;
 import it.unimi.dsi.fastutil.ints.IntIterators;
 import it.unimi.dsi.fastutil.ints.IntSet;
 import java.util.NoSuchElementException;
-import java.util.Spliterator;
-import java.util.Spliterators;
 import java.util.function.IntConsumer;
-import java.util.stream.IntStream;
-import java.util.stream.StreamSupport;
 
-abstract class AbstractNatBitSet extends AbstractIntSet implements NatBitSet {
-  static int SPLITERATOR_CHARACTERISTICS = Spliterator.ORDERED | Spliterator.SORTED
-      | Spliterator.DISTINCT | Spliterator.NONNULL
-      | Spliterator.SIZED;
-
-  protected static void checkNonNegative(int index) {
-    if (index < 0) {
-      throw new IndexOutOfBoundsException(String.format("Negative index %d ", index));
-    }
-  }
-
-  protected static void checkRange(int from, int to) {
-    if (to < from) {
-      throw new IndexOutOfBoundsException(String.format("From %d bigger than to %d", from, to));
-    }
-    if (from < 0) {
-      throw new IndexOutOfBoundsException(String.format("Negative from index %d ", from));
-    }
-  }
-
+public abstract class AbstractNatBitSet extends AbstractIntSet implements NatBitSet {
   @Override
   public boolean add(int index) {
     if (contains(index)) {
@@ -89,7 +66,7 @@ abstract class AbstractNatBitSet extends AbstractIntSet implements NatBitSet {
 
   @Override
   public void clearFrom(int from) {
-    checkNonNegative(from);
+    NatBitSetsUtil.checkNonNegative(from);
     clear(from, Integer.MAX_VALUE);
   }
 
@@ -109,11 +86,6 @@ abstract class AbstractNatBitSet extends AbstractIntSet implements NatBitSet {
       throw new NoSuchElementException();
     }
     return firstPresent;
-  }
-
-  @Override
-  public IntStream intStream() {
-    return StreamSupport.intStream(this::spliterator, SPLITERATOR_CHARACTERISTICS, false);
   }
 
   @Override
@@ -154,11 +126,6 @@ abstract class AbstractNatBitSet extends AbstractIntSet implements NatBitSet {
     }
     and(indices);
     return true;
-  }
-
-  @Override
-  public Spliterator.OfInt spliterator() {
-    return Spliterators.spliterator(iterator(), (long) size(), SPLITERATOR_CHARACTERISTICS);
   }
 
   @Override