+++ src/main/java/de/tum/in/naturals/set/SimpleBoundedNatBitSet.java
@@ -312,7 +312,7 @@ class SimpleBoundedNatBitSet extends AbstractBoundedNatBitSet {
   @Override
   public int nextAbsentIndex(int index) {
     assert checkConsistency();
-    checkNonNegative(index);
+    NatBitSetsUtil.checkNonNegative(index);
     if (index >= domainSize()) {
       return index;
     }
@@ -326,7 +326,7 @@ class SimpleBoundedNatBitSet extends AbstractBoundedNatBitSet {
   @Override
   public int nextPresentIndex(int index) {
     assert checkConsistency();
-    checkNonNegative(index);
+    NatBitSetsUtil.checkNonNegative(index);
     if (index >= domainSize()) {
       return -1;
     }
@@ -380,7 +380,7 @@ class SimpleBoundedNatBitSet extends AbstractBoundedNatBitSet {
   @Override
   public int previousAbsentIndex(int index) {
     assert checkConsistency();
-    checkNonNegative(index);
+    NatBitSetsUtil.checkNonNegative(index);
     if (index >= domainSize()) {
       return index;
     }
@@ -390,7 +390,7 @@ class SimpleBoundedNatBitSet extends AbstractBoundedNatBitSet {
   @Override
   public int previousPresentIndex(int index) {
     assert checkConsistency();
-    checkNonNegative(index);
+    NatBitSetsUtil.checkNonNegative(index);
     int clampedIndex = Math.min(index, domainSize() - 1);
     return complement
         ? bitSet.previousClearBit(clampedIndex)